Fix line number bug for XML text and comment nodes.
The DOM component has a bug when dealing with line numbers for text and comment nodes. This commit adds the property `fixLineNumbers` to the XMLDocument class. When setting this property to true, then the class fixes the line numbers when loading a document.

By the way, all XML related test cases and classes were refactored.